+++ scripts/config.mjs
@@ -7,8 +7,8 @@ export const CONDITION_EFFECT_DEHYDRATED = "dehydrated";
 export const CONDITION_EFFECT_MALNOURISHED = "malnourished";
 export const EXHAUSTION_PATH = "system.attributes.exhaustion";
 
-export const FOOD_NEEDS = { tiny: 0.25, sm: 1, med: 1, lg: 4, huge: 16, grg: 64 };
-export const WATER_NEEDS = { tiny: 0.25, sm: 1, med: 1, lg: 4, huge: 16, grg: 64 };
+export const FOOD_NEEDS = { tiny: 0.25, sm: 1, med: 1, lg: 4, huge: 16, grg: 64 }; // in pounds
+export const WATER_NEEDS = { tiny: 0.25, sm: 1, med: 1, lg: 4, huge: 16, grg: 64 }; // in gallons
 export const STARVATION_LIMIT = 5;
 
 export const MAGICAL_BERRIES_IDENTIFIER = "magical-berries";